initial refactor into separete components
