+++ src/user/UserCredentials.tsx
@@ -1,4 +1,4 @@
-import React, { Fragment, FunctionComponent, useMemo, useState } from "react";
+import React, { Fragment, useMemo, useState } from "react";
 import {
   AlertVariant,
   Button,
@@ -7,78 +7,31 @@ import {
   Dropdown,
   DropdownItem,
   DropdownPosition,
-  Form,
   FormGroup,
   KebabToggle,
-  Modal,
-  ModalVariant,
-  Select,
-  SelectOption,
-  SelectVariant,
-  Switch,
-  Text,
   TextInput,
-  TextVariants,
-  ValidatedOptions,
 } from "@patternfly/react-core";
-import {
-  Table,
-  TableBody,
-  TableComposable,
-  TableHeader,
-  TableVariant,
-  Td,
-  Th,
-  Thead,
-  Tr,
-} from "@patternfly/react-table";
+import { TableComposable, Td, Th, Thead, Tr } from "@patternfly/react-table";
 import { PencilAltIcon, CheckIcon, TimesIcon } from "@patternfly/react-icons";
 import type UserRepresentation from "@keycloak/keycloak-admin-client/lib/defs/userRepresentation";
 import { useTranslation } from "react-i18next";
-import { isEmpty } from "lodash/fp";
 import { useAlerts } from "../components/alert/Alerts";
 import { ListEmptyState } from "../components/list-empty-state/ListEmptyState";
 import { useAdminClient, useFetch } from "../context/auth/AdminClient";
 import { useWhoAmI } from "../context/whoami/WhoAmI";
-import { Controller, useForm, UseFormMethods, useWatch } from "react-hook-form";
-import { PasswordInput } from "../components/password-input/PasswordInput";
+import { useForm } from "react-hook-form";
 import { HelpItem } from "../components/help-enabler/HelpItem";
-import "./user-credentials.css";
 import { useConfirmDialog } from "../components/confirm-dialog/ConfirmDialog";
 import type CredentialRepresentation from "@keycloak/keycloak-admin-client/lib/defs/credentialRepresentation";
 import { FormAccess } from "../components/form-access/FormAccess";
-import { RequiredActionAlias } from "@keycloak/keycloak-admin-client/lib/defs/requiredActionProviderRepresentation";
-import { TimeSelector } from "../components/time-selector/TimeSelector";
+import { CredentialDataDialog } from "./user-credentials/CredentialDataDialog";
+import { ResetPasswordDialog } from "./user-credentials/ResetPasswordDialog";
+import { ResetCredentialDialog } from "./user-credentials/ResetCredentialDialog";
+
+import "./user-credentials.css";
 
 type UserCredentialsProps = {
   user: UserRepresentation;
-};
-
-type CredentialsForm = {
-  password: string;
-  passwordConfirmation: string;
-  temporaryPassword: boolean;
-};
-
-type CredentialResetForm = {
-  actions: RequiredActionAlias[];
-  lifespan: number;
-};
-
-const credFormDefaultValues: CredentialsForm = {
-  password: "",
-  passwordConfirmation: "",
-  temporaryPassword: true,
-};
-
-const credResetFormDefaultValues: CredentialResetForm = {
-  actions: [],
-  lifespan: 43200, // 12 hours
-};
-
-type DisplayDialogProps = {
-  titleKey: string;
-  onClose: () => void;
 };
 
 type UserLabelForm = {
@@ -95,154 +48,27 @@ type ExpandableCredentialRepresentation = {
   isExpanded: boolean;
 };
 
-const DisplayDialog: FunctionComponent<DisplayDialogProps> = ({
-  titleKey,
-  onClose,
-  children,
-}) => {
-  const { t } = useTranslation("users");
-  return (
-    <Modal
-      variant={ModalVariant.medium}
-      title={t(titleKey)}
-      isOpen={true}
-      onClose={onClose}
-    >
-      {children}
-    </Modal>
-  );
-};
-
-const CredentialsResetActionMultiSelect = (props: {
-  form: UseFormMethods<CredentialResetForm>;
-}) => {
-  const { t } = useTranslation("users");
-  const [open, setOpen] = useState(false);
-  const { form } = props;
-  const { control } = form;
-
-  return (
-    <FormGroup
-      label={t("resetActions")}
-      labelIcon={
-        <HelpItem
-          helpText="clients-help:resetActions"
-          fieldLabelId="resetActions"
-        />
-      }
-      fieldId="actions"
-    >
-      <Controller
-        name="actions"
-        defaultValue={[]}
-        control={control}
-        render={({ onChange, value }) => (
-          <Select
-            toggleId="actions"
-            variant={SelectVariant.typeaheadMulti}
-            chipGroupProps={{
-              numChips: 3,
-            }}
-            menuAppendTo="parent"
-            onToggle={(open) => setOpen(open)}
-            isOpen={open}
-            selections={value.map((o: string) => o)}
-            onSelect={(_, selectedValue) =>
-              onChange(
-                value.find((o: string) => o === selectedValue)
-                  ? value.filter((item: string) => item !== selectedValue)
-                  : [...value, selectedValue]
-              )
-            }
-            onClear={(event) => {
-              event.stopPropagation();
-              onChange([]);
-            }}
-            aria-label={t("resetActions")}
-          >
-            {Object.values(RequiredActionAlias).map((action, index) => (
-              <SelectOption
-                key={index}
-                value={action}
-                data-testid={`${action}-option`}
-              >
-                {t(action)}
-              </SelectOption>
-            ))}
-          </Select>
-        )}
-      />
-    </FormGroup>
-  );
-};
-
-const LifespanField = ({
-  form: { control },
-}: {
-  form: UseFormMethods<CredentialResetForm>;
-}) => {
-  const { t } = useTranslation("users");
-
-  return (
-    <FormGroup
-      fieldId="lifespan"
-      label={t("lifespan")}
-      isStack
-      labelIcon={
-        <HelpItem helpText="clients-help:lifespan" fieldLabelId="lifespan" />
-      }
-    >
-      <Controller
-        name="lifespan"
-        defaultValue={credResetFormDefaultValues.lifespan}
-        control={control}
-        render={({ onChange, value }) => (
-          <TimeSelector
-            value={value}
-            units={["minutes", "hours", "days"]}
-            onChange={onChange}
-            menuAppendTo="parent"
-          />
-        )}
-      />
-    </FormGroup>
-  );
-};
-
 export const UserCredentials = ({ user }: UserCredentialsProps) => {
   const { t } = useTranslation("users");
   const { whoAmI } = useWhoAmI();
   const { addAlert, addError } = useAlerts();
   const [key, setKey] = useState(0);
   const refresh = () => setKey(key + 1);
-  const [open, setOpen] = useState(false);
-  const [openSaveConfirm, setOpenSaveConfirm] = useState(false);
+  const [isOpen, setIsOpen] = useState(false);
   const [openCredentialReset, setOpenCredentialReset] = useState(false);
   const [kebabOpen, setKebabOpen] = useState({
     status: false,
     rowKey: "",
   });
   const adminClient = useAdminClient();
-  const form = useForm<CredentialsForm>({
-    defaultValues: credFormDefaultValues,
-  });
-  const resetForm = useForm<CredentialResetForm>({
-    defaultValues: credResetFormDefaultValues,
-  });
   const userLabelForm = useForm<UserLabelForm>({
     defaultValues: userLabelDefaultValues,
   });
-  const { control, errors, handleSubmit, register } = form;
-  const { control: resetControl, handleSubmit: resetHandleSubmit } = resetForm;
   const {
     getValues: getValues1,
     handleSubmit: handleSubmit1,
     register: register1,
   } = userLabelForm;
-  const [credentials, setCredentials] = useState<CredentialsForm>();
-  const [credentialsReset, setCredentialReset] = useState<CredentialResetForm>(
-    {} as CredentialResetForm
-  );
   const [userCredentials, setUserCredentials] = useState<
     CredentialRepresentation[]
   >([]);
@@ -289,107 +115,15 @@ export const UserCredentials = ({ user }: UserCredentialsProps) => {
     (credential) => credential.type === "password"
   );
 
-  const passwordWatcher = useWatch<CredentialsForm["password"]>({
-    control,
-    name: "password",
-  });
-
-  const resetActionWatcher = useWatch<CredentialResetForm["actions"]>({
-    control: resetControl,
-    name: "actions",
-  });
-
-  const passwordConfirmationWatcher = useWatch<
-    CredentialsForm["passwordConfirmation"]
-  >({
-    control,
-    name: "passwordConfirmation",
-  });
-
-  const isNotDisabled =
-    passwordWatcher !== "" && passwordConfirmationWatcher !== "";
-
-  const resetIsNotDisabled = !isEmpty(resetActionWatcher);
-
-  const toggleModal = () => {
-    setOpen(!open);
-  };
+  const toggleModal = () => setIsOpen(!isOpen);
 
   const toggleCredentialsResetModal = () => {
     setOpenCredentialReset(!openCredentialReset);
   };
 
-  const toggleConfirmSaveModal = () => {
-    setOpenSaveConfirm(!openSaveConfirm);
-  };
-
-  const saveUserPassword = async () => {
-    if (!credentials) {
-      return;
-    }
-
-    const passwordsMatch =
-      credentials.password === credentials.passwordConfirmation;
-
-    if (!passwordsMatch) {
-      addAlert(
-        isResetPassword
-          ? t("resetPasswordNotMatchError")
-          : t("savePasswordNotMatchError"),
-        AlertVariant.danger
-      );
-    } else {
-      try {
-        await adminClient.users.resetPassword({
-          id: user.id!,
-          credential: {
-            temporary: credentials.temporaryPassword,
-            type: "password",
-            value: credentials.password,
-          },
-        });
-        refresh();
-        addAlert(
-          isResetPassword
-            ? t("resetCredentialsSuccess")
-            : t("savePasswordSuccess"),
-          AlertVariant.success
-        );
-        setIsResetPassword(false);
-        setOpenSaveConfirm(false);
-      } catch (error) {
-        addError(
-          isResetPassword
-            ? "users:resetPasswordError"
-            : "users:savePasswordError",
-          error
-        );
-      }
-    }
-  };
-
-  const sendCredentialsResetEmail = async () => {
-    if (isEmpty(credentialsReset.actions)) {
-      return;
-    }
-
-    try {
-      await adminClient.users.executeActionsEmail({
-        id: user.id!,
-        actions: credentialsReset.actions,
-        lifespan: credentialsReset.lifespan,
-      });
-      refresh();
-      addAlert(t("credentialResetEmailSuccess"), AlertVariant.success);
-      setOpenCredentialReset(false);
-    } catch (error) {
-      addError(t("credentialResetEmailError"), error);
-    }
-  };
-
   const resetPassword = () => {
     setIsResetPassword(true);
-    setOpen(true);
+    toggleModal();
   };
 
   const [toggleDeleteDialog, DeleteConfirm] = useConfirmDialog({
@@ -464,233 +198,28 @@ export const UserCredentials = ({ user }: UserCredentialsProps) => {
 
   return (
     <>
-      {open && (
-        <Modal
-          variant={ModalVariant.small}
-          width={600}
-          title={
-            isResetPassword
-              ? t("resetPasswordFor", { username: user.username })
-              : t("setPasswordFor", { username: user.username })
-          }
-          isOpen
-          onClose={() => {
-            setIsResetPassword(false);
-            setOpen(false);
-          }}
-          actions={[
-            <Button
-              data-testid="okBtn"
-              key={`confirmBtn-${user.id}`}
-              variant="primary"
-              form="userCredentials-form"
-              onClick={() => {
-                setOpen(false);
-                setCredentials(form.getValues());
-                toggleConfirmSaveModal();
-              }}
-              isDisabled={!isNotDisabled}
-            >
-              {t("save")}
-            </Button>,
-            <Button
-              data-testid="cancel"
-              key={`cancelBtn-${user.id}`}
-              variant="link"
-              form="userCredentials-form"
-              onClick={() => {
-                setIsResetPassword(false);
-                setOpen(false);
-              }}
-            >
-              {t("cancel")}
-            </Button>,
-          ]}
-        >
-          <Form
-            id="userCredentials-form"
-            isHorizontal
-            className="keycloak__user-credentials__reset-form"
-          >
-            <FormGroup
-              name="password"
-              label={t("password")}
-              fieldId="password"
-              helperTextInvalid={t("common:required")}
-              validated={
-                errors.password
-                  ? ValidatedOptions.error
-                  : ValidatedOptions.default
-              }
-              isRequired
-            >
-              <PasswordInput
-                data-testid="passwordField"
-                name="password"
-                aria-label="password"
-                ref={register({ required: true })}
-              />
-            </FormGroup>
-            <FormGroup
-              name="passwordConfirmation"
-              label={
-                isResetPassword
-                  ? t("resetPasswordConfirmation")
-                  : t("passwordConfirmation")
-              }
-              fieldId="passwordConfirmation"
-              helperTextInvalid={t("common:required")}
-              validated={
-                errors.passwordConfirmation
-                  ? ValidatedOptions.error
-                  : ValidatedOptions.default
-              }
-              isRequired
-            >
-              <PasswordInput
-                data-testid="passwordConfirmationField"
-                name="passwordConfirmation"
-                aria-label="passwordConfirm"
-                ref={register({ required: true })}
-              />
-            </FormGroup>
-            <FormGroup
-              label={t("common:temporaryPassword")}
-              labelIcon={
-                <HelpItem
-                  helpText="temporaryPasswordHelpText"
-                  fieldLabelId="temporaryPassword"
-                />
-              }
-              fieldId="kc-temporaryPassword"
-            >
-              <Controller
-                name="temporaryPassword"
-                defaultValue={true}
-                control={control}
-                render={({ onChange, value }) => (
-                  <Switch
-                    className={"kc-temporaryPassword"}
-                    onChange={(value) => onChange(value)}
-                    isChecked={value}
-                    label={t("common:on")}
-                    labelOff={t("common:off")}
-                  />
-                )}
-              />
-            </FormGroup>
-          </Form>
-        </Modal>
-      )}
-      {openSaveConfirm && (
-        <Modal
-          variant={ModalVariant.small}
-          width={600}
-          title={
-            isResetPassword
-              ? t("resetPasswordConfirm")
-              : t("setPasswordConfirm")
-          }
-          isOpen
-          onClose={() => setOpenSaveConfirm(false)}
-          actions={[
-            <Button
-              data-testid="setPasswordBtn"
-              key={`confirmSaveBtn-${user.id}`}
-              variant="danger"
-              form="userCredentials-form"
-              onClick={() => {
-                handleSubmit(saveUserPassword)();
-              }}
-            >
-              {isResetPassword ? t("resetPassword") : t("savePassword")}
-            </Button>,
-            <Button
-              data-testid="cancel"
-              key={`cancelConfirmBtn-${user.id}`}
-              variant="link"
-              form="userCredentials-form"
-              onClick={() => {
-                setOpenSaveConfirm(false);
-              }}
-            >
-              {t("cancel")}
-            </Button>,
-          ]}
-        >
-          <Text component={TextVariants.h3}>
-            {isResetPassword
-              ? `${t("resetPasswordConfirmText")} ${user.username} ${t(
-                  "questionMark"
-                )}`
-              : `${t("setPasswordConfirmText")} ${user.username} ${t(
-                  "questionMark"
-                )}`}
-          </Text>
-        </Modal>
+      {isOpen && (
+        <ResetPasswordDialog
+          user={user}
+          isResetPassword={isResetPassword}
+          onClose={() => setIsOpen(false)}
+        />
       )}
       {openCredentialReset && (
-        <Modal
-          variant={ModalVariant.medium}
-          title={t("credentialReset")}
-          isOpen
-          onClose={() => {
-            setOpenCredentialReset(false);
-          }}
-          data-testid="credential-reset-modal"
-          actions={[
-            <Button
-              data-testid="okBtn"
-              key={`confirmBtn-${user.id}`}
-              variant="primary"
-              form="userCredentialsReset-form"
-              onClick={() => {
-                setCredentialReset(resetForm.getValues());
-                resetHandleSubmit(sendCredentialsResetEmail)();
-              }}
-              isDisabled={!resetIsNotDisabled}
-            >
-              {t("credentialResetConfirm")}
-            </Button>,
-            <Button
-              data-testid="cancel"
-              key={`cancelBtn-${user.id}`}
-              variant="link"
-              form="userCredentialsReset-form"
-              onClick={() => {
-                setOpenCredentialReset(false);
-              }}
-            >
-              {t("cancel")}
-            </Button>,
-          ]}
-        >
-          <Form id="userCredentialsReset-form" isHorizontal>
-            <CredentialsResetActionMultiSelect form={resetForm} />
-            <LifespanField form={resetForm} />
-          </Form>
-        </Modal>
+        <ResetCredentialDialog
+          userId={user.id!}
+          onClose={() => setOpenCredentialReset(false)}
+        />
       )}
       <DeleteConfirm />
       {showData && Object.keys(selectedCredential).length !== 0 && (
-        <DisplayDialog
-          titleKey={t("passwordDataTitle")}
+        <CredentialDataDialog
+          credentialData={rows}
           onClose={() => {
             setShowData(false);
             setSelectedCredential({});
           }}
-        >
-          <Table
-            aria-label="password-data"
-            data-testid="password-data-dialog"
-            variant={TableVariant.compact}
-            cells={[t("showPasswordDataName"), t("showPasswordDataValue")]}
-            rows={rows}
-          >
-            <TableHeader />
-            <TableBody />
-          </Table>
-        </DisplayDialog>
+        />
       )}
       {userCredentials.length !== 0 && passwordTypeFinder === undefined && (
         <>
@@ -701,7 +230,7 @@ export const UserCredentials = ({ user }: UserCredentialsProps) => {
             variant="primary"
             form="userCredentials-form"
             onClick={() => {
-              setOpen(true);
+              setIsOpen(true);
             }}
           >
             {t("savePassword")}
